Move TerminalProcessManager, node->browser
