+++ src/tsconfig.strictNullChecks.json
@@ -285,6 +285,7 @@
 		"./vs/workbench/contrib/terminal/browser/terminalInstance.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalLinkHandler.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalPanel.ts",
+		"./vs/workbench/contrib/terminal/browser/terminalProcessManager.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalQuickOpen.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalTab.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalWidgetManager.ts",
@@ -297,7 +298,6 @@
 		"./vs/workbench/contrib/terminal/common/terminalService.ts",
 		"./vs/workbench/contrib/terminal/node/terminal.ts",
 		"./vs/workbench/contrib/terminal/node/terminalProcess.ts",
-		"./vs/workbench/contrib/terminal/node/terminalProcessManager.ts",
 		"./vs/workbench/contrib/terminal/node/windowsShellHelper.ts",
 		"./vs/workbench/contrib/terminal/test/electron-browser/terminalColorRegistry.test.ts",
 		"./vs/workbench/contrib/terminal/test/electron-browser/terminalCommandTracker.test.ts",
