+++ src/vs/base/common/processes.ts
@@ -3,6 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { IProcessEnvironment } from 'vs/base/common/platform';
+
 /**
  * Options to be passed to the external program or shell.
  */
@@ -83,4 +85,31 @@ export const enum TerminateResponseCode {
 	Unknown = 1,
 	AccessDenied = 2,
 	ProcessNotFound = 3,
+}
+
+/**
+ * Sanitizes a VS Code process environment by removing all Electron/VS Code-related values.
+ */
+export function sanitizeProcessEnvironment(env: IProcessEnvironment, ...preserve: string[]): void {
+	const set = preserve.reduce((set, key) => {
+		set[key] = true;
+		return set;
+	}, {} as Record<string, boolean>);
+	const keysToRemove = [
+		/^ELECTRON_.+$/,
+		/^GOOGLE_API_KEY$/,
+		/^VSCODE_.+$/,
+		/^SNAP(|_.*)$/
+	];
+	const envKeys = Object.keys(env);
+	envKeys
+		.filter(key => !set[key])
+		.forEach(envKey => {
+			for (let i = 0; i < keysToRemove.length; i++) {
+				if (envKey.search(keysToRemove[i]) !== -1) {
+					delete env[envKey];
+					break;
+				}
+			}
+		});
 }
